+++ src/features/Login/LoginPresenter.tsx
@@ -1,8 +1,9 @@
-import { Box, Typography, Container, Button, Divider, Snackbar, Alert } from '@mui/material';
+import { Box, Typography, Container, Button, Divider } from '@mui/material';
 import GoogleSignInButton from '../../components/buttonicon/GoogleSignInButton';
 import { useAuth } from '../../providers/auth';
-import { Link as RouterLink, useLocation, useNavigate  } from 'react-router';
-import React, { useEffect, useState, useCallback } from 'react'
+import { Link as RouterLink } from 'react-router';
+import React from 'react'
+import ToastSnackbar from '../../components/toastSnackbar'
 
 export const LoginPresenter = () => {
   const authBaseUrl = import.meta.env.VITE_AUTH_BASE_URL;
@@ -12,54 +13,6 @@ export const LoginPresenter = () => {
     e.preventDefault();
     window.location.href = `${authBaseUrl}/auth/google_oauth2`;
   };
-  const location = useLocation()
-  const navigate = useNavigate()
-
-  const [toastOpen, setToastOpen] = useState(false)
-  const [toastMessage, setToastMessage] = useState('')
-  const [toastSeverity, setToastSeverity] = useState<'error' | 'warning' | 'info' | 'success'>('info') // MUIのAlertの種類
-
-  const handleToastClose = useCallback((_event?: unknown, reason?: string) => {
-    if (reason === 'clickaway') return
-    setToastOpen(false)
-  }, [])
-
-  useEffect(() => {
-    const params = new URLSearchParams(location.search)
-    const msg = params.get('message')
-    const level = params.get('messageLevel')
-
-    if (msg) {
-      // デコード（Rails 側で自動エンコードされている想定）
-      const decoded = decodeURIComponent(msg)
-      setToastMessage(decoded)
-
-      switch ((level || '').toLowerCase()) {
-        case 'error':
-          setToastSeverity('error')
-          break
-        case 'warning':
-          setToastSeverity('warning')
-          break
-        case 'success':
-          setToastSeverity('success')
-          break
-        default:
-          setToastSeverity('info')
-      }
-
-      setToastOpen(true)
-
-      // クエリを消してリロード時に再表示されないようにする
-      // react-router の navigate がある場合は replace、なければ history API を使う
-      try {
-        navigate(location.pathname + location.hash, { replace: true })
-      } catch {
-        window.history.replaceState({}, document.title, location.pathname + location.hash)
-      }
-    }
-  // location.search が変わったら再評価
-  }, [location.search, location.pathname, location.hash, navigate])
 
   return (
     <Container maxWidth='sm'>
@@ -115,26 +68,7 @@ export const LoginPresenter = () => {
               <GoogleSignInButton onClick={handleGoogleAuth} className="" />
             </>
           )}
-          <Snackbar
-            open={toastOpen}
-            autoHideDuration={5000}
-            onClose={handleToastClose}
-            anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-            sx={{
-              mb: 12, // 下の余白
-              zIndex: (theme) => theme.zIndex.snackbar + 1000, // 前面に出す
-            }}
-          >
-            <Alert
-              onClose={handleToastClose}
-              severity={toastSeverity}
-              sx={(theme) => ({
-                width: '100%',
-                boxShadow: theme.shadows[6],
-              })}>
-              {toastMessage}
-            </Alert>
-          </Snackbar>
+          <ToastSnackbar />
         </>
       </Box>
     </Container>
